+++ Game/src/laser/lasermanager.cpp
@@ -53,7 +53,7 @@ void LaserManager::Update( float delta )
 
     using namespace Genesis;
 
-    const unsigned int lasersCount = m_Lasers.size();
+    const size_t lasersCount = m_Lasers.size();
     if ( lasersCount == 0 )
     {
         return;
@@ -105,13 +105,13 @@ void LaserManager::Render()
 {
     using namespace Genesis;
 
-    const unsigned int lasersCount = m_Lasers.size();
+    const size_t lasersCount = m_Lasers.size();
     if ( lasersCount == 0 )
     {
         return;
     }
 
-    const unsigned int verticesCount = lasersCount * 6;
+    const size_t verticesCount = lasersCount * 6;
 
     RenderSystem* pRenderSystem = FrameWork::GetRenderSystem();
     pRenderSystem->SetBlendMode( BlendMode::Add );
@@ -119,10 +119,10 @@ void LaserManager::Render()
     m_pShader->Use();
 
     pRenderSystem->SetRenderTarget( RenderTargetId::Glow );
-    m_pVertexBuffer->Draw( verticesCount );
+    m_pVertexBuffer->Draw( static_cast<uint32_t>( verticesCount ) );
 
     pRenderSystem->SetRenderTarget( RenderTargetId::Default );
-    m_pVertexBuffer->Draw( verticesCount );
+    m_pVertexBuffer->Draw( static_cast<uint32_t>( verticesCount ) );
 
     pRenderSystem->SetBlendMode( BlendMode::Disabled );
 
@@ -131,7 +131,7 @@ void LaserManager::Render()
 
 void LaserManager::AddLaser( const Laser& laser )
 {
-    const int numLasers = m_Lasers.size();
+    const size_t numLasers = m_Lasers.size();
     if ( numLasers == sLaserManagerCapacity )
     {
         Genesis::FrameWork::GetLogger()->LogWarning( "Unable to add laser, manager at capacity (%d)", sLaserManagerCapacity );
