+++ Game/src/particles/particleemitter.cpp
@@ -97,8 +97,8 @@ void ParticleEmitter::CreateParticle()
 
 Particle* ParticleEmitter::GetAvailableParticle()
 {
-    int numEmitters = m_Particles.size();
-    for ( int i = m_Idx; i < numEmitters; ++i )
+    size_t numEmitters = m_Particles.size();
+    for ( size_t i = m_Idx; i < numEmitters; ++i )
     {
         if ( m_Particles[ i ].IsAlive() == false )
         {
@@ -108,7 +108,7 @@ Particle* ParticleEmitter::GetAvailableParticle()
         }
     }
 
-    for ( int i = 0; i < m_Idx; ++i )
+    for ( size_t i = 0; i < m_Idx; ++i )
     {
         if ( m_Particles[ i ].IsAlive() == false )
         {
