+++ Game/src/trail/trailmanagerrep.h
@@ -49,7 +49,7 @@ private:
     TrailManager* m_pTrailManager;
     Genesis::Shader* m_pShader;
     Genesis::VertexBuffer* m_pVertexBuffer;
-    unsigned int m_NumVertices;
+    size_t m_NumVertices;
 };
 
 } // namespace Hexterminate
